Optimizing KeyExpr processing
When processing the internal string representation of a kotlin KeyExpr instance
through JNI, a rust KeyExpr instance was generated through the "try_from" function.
This is unnecessary since the string representation was already validated during the
construction phase of the KeyExpr object on Kotlin.
Therefore, this commit introduces an optimization in which the rust key expression
is created with "from_string_unchecked", since we already guarantee that the string
representation provided is valid and has already been checked.